Extract Message from MessageList into a component
